requirements.txt
mcp[cli]>=1.0.0
python-roborock>=2.0.0
